% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgl_navigate.R
\name{best_model.msgl}
\alias{best_model.msgl}
\title{Index of best model}
\usage{
\method{best_model}{msgl}(object, ...)
}
\arguments{
\item{object}{a msgl object}

\item{...}{additional parameters (ignored)}
}
\value{
index of the best model.
}
\description{
Returns the index of the best model, in terms of lowest error rate
}
\author{
Martin Vincent
}

