% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{msgl.c.config}
\alias{msgl.c.config}
\title{Featch information about the C side configuration of the package}
\usage{
msgl.c.config()
}
\value{
list
}
\description{
Featch information about the C side configuration of the package
}
\author{
Martin Vicnet
}

