% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv.R
\name{msgl.cv}
\alias{msgl.cv}
\title{Deprecated cv function}
\usage{
msgl.cv(x, classes, sampleWeights = NULL, grouping = NULL,
  groupWeights = NULL, parameterWeights = NULL, alpha = 0.5,
  standardize = TRUE, lambda, d = 100, fold = 10L, cv.indices = list(),
  intercept = TRUE, sparse.data = is(x, "sparseMatrix"),
  max.threads = NULL, use_parallel = FALSE,
  algorithm.config = msgl.standard.config)
}
\description{
Deprecated cv function
}
\keyword{internal}
