% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{msgl.lambda.seq}
\alias{msgl.lambda.seq}
\title{Deprecated lambda function}
\usage{
msgl.lambda.seq(x, classes, sampleWeights = NULL, grouping = NULL,
  groupWeights = NULL, parameterWeights = NULL, alpha = 0.5,
  d = 100L, standardize = TRUE, lambda.min, intercept = TRUE,
  sparse.data = is(x, "sparseMatrix"), lambda.min.rel = FALSE,
  algorithm.config = msgl.standard.config)
}
\description{
Deprecated lambda function
}
\keyword{internal}
