% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{msgl}
\alias{msgl}
\title{Deprecated fit function}
\usage{
msgl(x, classes, sampleWeights = NULL, grouping = NULL,
  groupWeights = NULL, parameterWeights = NULL, alpha = 0.5,
  standardize = TRUE, lambda, d = 100, return_indices = NULL,
  intercept = TRUE, sparse.data = is(x, "sparseMatrix"),
  algorithm.config = msgl.standard.config)
}
\description{
Deprecated fit function
}
\keyword{internal}
