% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\docType{data}
\name{msgl.standard.config}
\alias{msgl.standard.config}
\title{Standard msgl algorithm configuration}
\format{An object of class \code{list} of length 15.}
\usage{
msgl.standard.config
}
\description{
\code{msgl.standard.config <- msgl.algorithm.config()}
}
\author{
Martin Vincent
}
\keyword{datasets}
