% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subsampling.R
\name{msgl.subsampling}
\alias{msgl.subsampling}
\title{Deprecated subsampling function}
\usage{
msgl.subsampling(x, classes, sampleWeights = NULL, grouping = NULL,
  groupWeights = NULL, parameterWeights = NULL, alpha = 0.5,
  standardize = TRUE, lambda, d = 100, training, test,
  intercept = TRUE, sparse.data = is(x, "sparseMatrix"),
  collapse = FALSE, max.threads = NULL, use_parallel = FALSE,
  algorithm.config = msgl.standard.config)
}
\description{
Deprecated subsampling function
}
\keyword{internal}
