% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lambda.R
\name{msgl_dense_sgl_lambda_R}
\alias{msgl_dense_sgl_lambda_R}
\title{C interface}
\usage{
msgl_dense_sgl_lambda_R(data, block_dim, groupWeights, parameterWeights, alpha,
  d, lambda.min, lambda.min.rel, algorithm.config)
}
\description{
C interface
}
\keyword{internal}
