% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{msgl_sparse_sgl_fit_R}
\alias{msgl_sparse_sgl_fit_R}
\title{C interface}
\usage{
msgl_sparse_sgl_fit_R(data, block_dim, groupWeights, parameterWeights, alpha,
  lambda, idx, algorithm.config)
}
\description{
C interface
}
\keyword{internal}
