\docType{data}
\name{sgl.standard.config}
\alias{sgl.standard.config}
\title{Standard algorithm configuartion}
\format{List of 21
 $ tolerance_penalized_main_equation_loop      : num 1e-10
 $ tolerance_penalized_inner_loop_alpha        : num 1e-04
 $ tolerance_penalized_inner_loop_beta         : num 5
 $ tolerance_penalized_middel_loop_alpha       : num 0.01
 $ tolerance_penalized_middel_loop_beta        : num 0
 $ tolerance_penalized_outer_loop_alpha        : num 0.01
 $ tolerance_penalized_outer_loop_beta         : num 0
 $ tolerance_penalized_outer_loop_gamma        : num 1e-05
 $ general_tolerance_unpenalized               : num 0.001
 $ use_bound_optimization                      : logi TRUE
 $ use_active_set_optimization                 : logi FALSE
 $ active_set_threshold_alpha                  : num 0.5
 $ active_set_threshold_beta                   : num 0
 $ use_stepsize_optimization_in_penalizeed_loop: logi TRUE
 $ stepsize_opt_penalized_initial_t            : num 1
 $ stepsize_opt_penalized_a                    : num 0.1
 $ stepsize_opt_penalized_b                    : num 0.5
 $ stepsize_opt_unpenalized_initial_t          : num 1
 $ stepsize_opt_unpenalized_a                  : num 0.1
 $ stepsize_opt_unpenalized_b                  : num 0.5
 $ verbose                                     : logi FALSE}
\usage{
  sgl.standard.config
}
\description{
  Standard algorithm configuartion
}
\author{
  Martin Vicnet
}
\keyword{datasets}

