% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{sim.data}
\alias{sim.data}
\title{Simulated data set}
\description{
The use of this data set is only intended for testing and examples.
The data set contains 100 simulated samples grouped into 10 classes.
For each sample 400 features have been simulated.
}
\keyword{data}

