% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgl.R
\docType{data}
\name{SimData}
\alias{SimData}
\title{Simulated data set}
\format{A design matrix and a class vector
\describe{
  \item{x}{design matrix}
  \item{classes}{class vector}
  ...
}}
\description{
The use of this data set is only intended for testing and examples.
The data set contains 100 simulated samples grouped into 10 classes.
For each sample 400 features have been simulated.
}
\keyword{data}

