\name{models.msgl}
\alias{models.msgl}
\title{Exstract the estimated models}
\usage{
  \method{models}{msgl} (x, index = 1:nmod(x), ...)
}
\arguments{
  \item{x}{a msgl object}

  \item{index}{the models to be returned}
}
\value{
  a list of sparse matrices
}
\description{
  Exstract the estimated models
}
\author{
  Martin Vincent
}

