% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/msgl.R
\docType{data}
\name{PrimaryCancers}
\alias{PrimaryCancers}
\title{Primary cancer samples.}
\format{A design matrix and a class vector
\describe{
  \item{x}{design matrix}
  \item{classes}{class vector}
}}
\description{
Data set consisting of microRNA normalized expression measurements of primary cancer samples.
}
\references{
\url{http://www.ncbi.nlm.nih.gov/pubmed/24463184}
}
\keyword{data}

